Hyprland: Remove Waylock module
I'm removing this module because I was unable to integrate it with
Hyprland correctly and I don't want to allocate the time to fix that.

I dislike this decision because I prefer Waylock for it's simplicity.

I will be focusing my efforts on getting it to work with Hyprlock in the
short term.